Replace old sidebar with new one, remove Prompt from topbar
