+++ fireaid_web/src/app/library/page.tsx
@@ -36,7 +36,7 @@ export default function LibraryPage() {
 
   return (
     <div className="grid min-w-0 grid-cols-1 gap-6 xl:grid-cols-[288px_minmax(0,1fr)]">
-      <FireAIDSidebar active="library" />
+      <FireAIDSidebar/>
 
       <div className="min-w-0 space-y-4">
         {/* Header */}
