+++ fireaid_web/src/app/portal/page.tsx
@@ -4,15 +4,12 @@ import { useState, useEffect, useRef } from "react";
 import Link from "next/link";
 import Topbar from "@/components/topbar/Topbar";
 import {
-  LayoutGrid,
-  Database,
-  Terminal,
-  BarChart3,
   Send,
   Bot,
   User,
   RefreshCw,
 } from "lucide-react";
+import FireAIDSidebar from "@/components/layout/FireAIDSidebar";
 
 type Message = { role: "user" | "assistant"; content: string; time: string };
 
@@ -24,11 +21,6 @@ const SUGGESTED = [
   "What fuel types cause large fires?",
 ];
 
-const SIDEBAR_ITEMS = [
-  { key: "apps",          label: "APPS",   icon: LayoutGrid, href: "/apps" },
-  { key: "data",          label: "DATA",   icon: Database,   href: "/data" },
-  { key: "prompt",        label: "PROMPT", icon: Terminal,   href: "/prompt" },
-];
 
 export default function PortalPage() {
   const [messages, setMessages] = useState<Message[]>([]);
@@ -79,18 +71,7 @@ export default function PortalPage() {
 
       <div className="flex flex-1">
         {/* SIDEBAR */}
-        <aside className="w-14 bg-white border-r border-slate-200 flex flex-col items-center py-6 gap-2 shrink-0">
-          {SIDEBAR_ITEMS.map(({ key, label, icon: Icon, href }) => (
-            <Link
-              key={key}
-              href={href}
-              className="w-11 py-2.5 rounded-xl flex flex-col items-center gap-1 text-slate-400 hover:bg-slate-100 hover:text-[#003366] transition text-[8px] font-semibold tracking-widest"
-            >
-              <Icon size={18} strokeWidth={1.8} />
-              {label}
-            </Link>
-          ))}
-        </aside>
+        <FireAIDSidebar/>
 
         {/* MAIN */}
         <main className="flex-1 flex flex-col bg-slate-50">
