+++ fireaid_web/src/app/prediction/page.tsx
@@ -85,7 +85,7 @@ export default function PredictionPage() {
 
   return (
     <div className="grid min-w-0 grid-cols-1 gap-6 xl:grid-cols-[288px_minmax(0,1fr)]">
-      <FireAIDSidebar active="prediction" />
+      <FireAIDSidebar/>
 
       <div className="min-w-0 space-y-4">
         {/* Header */}
