+++ fireaid_web/src/app/projects/page.tsx
@@ -2,13 +2,8 @@
 
 import { useState, useEffect } from "react";
 import Link from "next/link";
-import { Search, Plus, Trash2, LayoutGrid, Database, BarChart3, Terminal, FolderOpen } from "lucide-react";
-
-const SIDEBAR_ITEMS = [
-  { key: "apps",          label: "APPS",   icon: LayoutGrid, href: "/apps" },
-  { key: "data",          label: "DATA",   icon: Database,   href: "/data" },
-  { key: "prompt",        label: "PROMPT", icon: Terminal,   href: "/prompt" },
-];
+import { Search, Plus, Trash2, FolderOpen } from "lucide-react";
+import FireAIDSidebar from "@/components/layout/FireAIDSidebar";
 
 type Project = {
   _id: string;
@@ -78,15 +73,7 @@ export default function ProjectsPage() {
   return (
     <div className="flex w-full min-h-screen">
       {/* SIDEBAR */}
-      <aside className="w-14 bg-white border-r border-slate-200 flex flex-col items-center py-6 gap-2 shrink-0">
-        {SIDEBAR_ITEMS.map(({ key, label, icon: Icon, href }) => (
-          <Link key={key} href={href}
-            className="rounded-xl text-slate-400 hover:bg-slate-100 hover:text-[#003366] transition w-11 py-2.5 flex flex-col items-center gap-1 text-[8px] font-semibold tracking-widest">
-            <Icon size={18} strokeWidth={1.8} />
-            {label}
-          </Link>
-        ))}
-      </aside>
+      <FireAIDSidebar/>
 
       {/* MAIN */}
       <main className="flex-1 flex flex-col bg-slate-50">
