+++ fireaid_web/src/app/team/page.tsx
@@ -1,12 +1,8 @@
 import Link from "next/link";
 import { LayoutGrid, Database, Terminal } from "lucide-react";
 import Topbar from "@/components/topbar/Topbar";
+import FireAIDSidebar from "@/components/layout/FireAIDSidebar";
 
-const SIDEBAR_ITEMS = [
-  { key: "apps",   label: "APPS",   icon: LayoutGrid, href: "/apps" },
-  { key: "data",   label: "DATA",   icon: Database,   href: "/data" },
-  { key: "prompt", label: "PROMPT", icon: Terminal,   href: "/prompt" },
-];
 
 const TEAM = [
     
@@ -22,15 +18,7 @@ const TEAM = [
 export default function TeamPage() {
   return (
     <div className="flex w-full min-h-screen">
-      <aside className="w-14 bg-white border-r border-slate-200 flex flex-col items-center py-6 gap-2 shrink-0">
-        {SIDEBAR_ITEMS.map(({ key, label, icon: Icon, href }) => (
-          <Link key={key} href={href}
-            className="rounded-xl text-slate-400 hover:bg-slate-100 hover:text-[#003366] transition w-11 py-2.5 flex flex-col items-center gap-1 text-[8px] font-semibold tracking-widest">
-            <Icon size={18} strokeWidth={1.8} />
-            {label}
-          </Link>
-        ))}
-      </aside>
+      <FireAIDSidebar/>
 
       <main className="flex-1 flex flex-col">
         <div className="bg-slate-50 px-12 py-10 border-b border-slate-200">
