+++ fireaid_web/src/components/topbar/Topbar.tsx
@@ -41,7 +41,6 @@ export default function Topbar() {
             { label: "Chat",   href: "/portal" },
             { label: "Apps",   href: "/apps" },
             { label: "Data",   href: "/data" },
-            { label: "Prompt", href: "/prompt" },
             { label: "Team",   href: "/team" },
         
           ].map(({ label, href }) => (
